+++ AcuDragon/ApiClient.swift
@@ -1,0 +1,56 @@
+//
+//  ApiClient.swift
+//  AcuDragon
+//
+//  Created by Emiko Clark on 4/5/18.
+//  Copyright © 2018 Emiko Clark. All rights reserved.
+//
+
+import UIKit
+
+class ApiClient {
+    
+    static func fetchVideos( completion: @escaping(Video)->() ) {
+        var videoObject = Video()
+        var videoItemsArr = [Items]()
+        
+        let requestString = "https://www.googleapis.com/youtube/v3/search?key=\(myAPIKey)&channelId=UCD5kT8GTKnbYl9WxgnLM0aA&part=snippet,id"
+        let urlRequest =  URL(string: requestString)
+        
+        URLSession.shared.dataTask(with: urlRequest!) { (data, response, error) in
+            if error != nil {
+                print(error!)
+                return
+            }
+            guard let data  = data else { return }
+            
+            print(data.description)
+            
+            do {
+                let json = try JSONDecoder().decode(Video.self, from: data)
+                
+                guard let etag = json.etag else { print("mainJson etag nil"); return }
+                guard let itemsArray = json.items else { print("error creating json"); return }
+                
+                videoObject.etag = etag
+                
+                for vid in itemsArray {
+                    videoItemsArr.append(vid)
+                }
+                videoObject.items?.append(contentsOf: videoItemsArr)
+                
+                DispatchQueue.main.async() {
+                    completion(videoObject)
+                }
+            } catch let error {
+                print(error)
+            }
+        }.resume()
+    }
+    
+    
+    static func downloadImage(urlString: String, completion: @escaping(UIImage)->()) {
+        
+    }
+    
+}
